LayoutTests/imported/w3c/web-platform-tests/navigation-api/navigation-history-entry/no-referrer-url-censored-expected.txt



FAIL The url of a document with no-referrer referrer policy is censored in NavigationHistoryEntry assert_equals: expected 3 but got 1
